// src/api/generated/models/PositionResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * eGovFrame 5.0 API
 * API Documentation for eGovFrame 5.0 Application
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface PositionResponse
 */
export interface PositionResponse {
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    positionId?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    positionName?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    positionCode?: string;
    /**
     * 
     * @type {number}
     * @memberof PositionResponse
     */
    positionLevel?: number;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    positionDescription?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    tenantId?: string;
    /**
     * 
     * @type {string}
     * @memberof PositionResponse
     */
    useAt?: string;
    /**
     * 
     * @type {Date}
     * @memberof PositionResponse
     */
    createdDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof PositionResponse
     */
    updatedDate?: Date;
}

/**
 * Check if a given object implements the PositionResponse interface.
 */
export function instanceOfPositionResponse(value: object): value is PositionResponse {
    return true;
}

export function PositionResponseFromJSON(json: any): PositionResponse {
    return PositionResponseFromJSONTyped(json, false);
}

export function PositionResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PositionResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'positionId': json['positionId'] == null ? undefined : json['positionId'],
        'positionName': json['positionName'] == null ? undefined : json['positionName'],
        'positionCode': json['positionCode'] == null ? undefined : json['positionCode'],
        'positionLevel': json['positionLevel'] == null ? undefined : json['positionLevel'],
        'positionDescription': json['positionDescription'] == null ? undefined : json['positionDescription'],
        'tenantId': json['tenantId'] == null ? undefined : json['tenantId'],
        'useAt': json['useAt'] == null ? undefined : json['useAt'],
        'createdDate': json['createdDate'] == null ? undefined : (new Date(json['createdDate'])),
        'updatedDate': json['updatedDate'] == null ? undefined : (new Date(json['updatedDate'])),
    };
}

export function PositionResponseToJSON(json: any): PositionResponse {
    return PositionResponseToJSONTyped(json, false);
}

export function PositionResponseToJSONTyped(value?: PositionResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'positionId': value['positionId'],
        'positionName': value['positionName'],
        'positionCode': value['positionCode'],
        'positionLevel': value['positionLevel'],
        'positionDescription': value['positionDescription'],
        'tenantId': value['tenantId'],
        'useAt': value['useAt'],
        'createdDate': value['createdDate'] == null ? undefined : ((value['createdDate']).toISOString()),
        'updatedDate': value['updatedDate'] == null ? undefined : ((value['updatedDate']).toISOString()),
    };
}